chore: style and move components
